更改目录结构和代码结构，方便其他项目可以引入
