+++ packages/api.go
@@ -2,30 +2,42 @@
  * @Author: ph4nt0mer
  * @Date: 2022-09-01 15:36:10
  * @LastEditors: rootphantomer
- * @LastEditTime: 2022-09-15 10:16:42
- * @FilePath: /quake_go/src/apis/api.go
+ * @LastEditTime: 2022-09-19 10:39:10
+ * @FilePath: /quake_go/packages/api.go
  * @Description:封装请求接口
  *
  * Copyright (c) 2022 by ph4nt0mer, All Rights Reserved.
  */
-package apis
+package packages
 
 import (
 	"encoding/json"
 	"fmt"
 	. "quake/src/model"
 	"quake/src/setting"
-	"quake/src/tools"
 	"quake/src/utils"
 	"strconv"
 	"strings"
+
+	. "time"
 )
+
+type Reqjson struct {
+	Query        string `json:"query"`
+	Start        string `json:"start,omitempty"`
+	Size         string `json:"size,omitempty"`
+	Ignore_cache bool   `json:"ignore_cache,omitempty"`
+	Start_time   Time   `json:"start_time,omitempty"`
+	End_time     Time   `json:"end_time,omitempty"`
+	Field        string `json:"-"`
+	Query_txt    string `json:"-"`
+}
 
 func FilterableServiceGET(token string) {
 	// 获取服务数据筛选字段
 	// curl -X GET "https://quake.360.cn/api/v3/filterable/field/quake_service" -H "X-QuakeToken: d17140ae-xxxx-xxx-xxxx-c0818b2bbxxx"
 	uri := "/filterable/field/quake_service"
-	tools.Apis(setting.URL+uri, "GET", []byte{}, token)
+	Apis(setting.URL+uri, "GET", []byte{}, token)
 }
 func SearchServicePost(reqjson Reqjson, token string) {
 	// 服务数据实时查询接口
@@ -64,7 +76,7 @@ func SearchServicePost(reqjson Reqjson, token string) {
 	}
 	fmt.Println("->", string(setting.URL+uri))
 	fmt.Println(string(datajson))
-	body := tools.Apis(setting.URL+uri, "POST", datajson, token)
+	body := Apis(setting.URL+uri, "POST", datajson, token)
 	data_result := utils.RespLoadJson[SearchJson](body).Data
 	if reqjson.Field != "" && reqjson.Field != "ip,port" {
 		for index, value := range data_result {
@@ -101,25 +113,25 @@ func ScrollServicePost(query []byte, start string, size string, token string) {
 	//     "end_time": "2021-05-20 01:13:14"
 	// }'
 	uri := "/scroll/quake_service"
-	tools.Apis(setting.URL+uri, "POST", query, token)
+	Apis(setting.URL+uri, "POST", query, token)
 }
 func AggregationServiceGet(token string) {
 	// 获取聚合数据筛选字段
 	// curl -X GET "https://quake.360.cn/api/v3/aggregation/quake_service" -H "X-QuakeToken: d17140ae-xxxx-xxx-xxxx-c0818b2bbxxx"
 	uri := "/aggregation/quake_service"
-	tools.Apis(setting.URL+uri, "GET", []byte{}, token)
+	Apis(setting.URL+uri, "GET", []byte{}, token)
 }
 func AggregationServicePost(query []byte, start string, size string, token string) {
 	// 获取聚合数据筛选字段
 	// curl -X GET "https://quake.360.cn/api/v3/aggregation/quake_service" -H "X-QuakeToken: d17140ae-xxxx-xxx-xxxx-c0818b2bbxxx"
 	uri := "/aggregation/quake_service"
-	tools.Apis(setting.URL+uri, "POST", query, token)
+	Apis(setting.URL+uri, "POST", query, token)
 }
 
 func InfoGet(token string) {
 	// 个人信息接口
 	uri := "/user/info"
-	info := tools.Apis(setting.URL+uri, "GET", []byte{}, token)
+	info := Apis(setting.URL+uri, "GET", []byte{}, token)
 	data_result, user_result := utils.InfoLoadJson(info)
 	fmt.Println("#用户名:", user_result["username"])
 	fmt.Println("#邮  箱:", user_result["email"])
@@ -131,7 +143,7 @@ func InfoGet(token string) {
 }
 func FaviconPost(query string, token string) {
 	uri := "/query/similar_icon/aggregation"
-	tools.Apis(setting.URL+uri, "GET", []byte{}, token)
+	Apis(setting.URL+uri, "GET", []byte{}, token)
 }
 
 func HostSearchPost(reqjson Reqjson, token string) {
@@ -172,7 +184,7 @@ func HostSearchPost(reqjson Reqjson, token string) {
 	}
 	fmt.Println("->", string(setting.URL+uri))
 	fmt.Println(string(datajson))
-	body := tools.Apis(setting.URL+uri, "POST", datajson, token)
+	body := Apis(setting.URL+uri, "POST", datajson, token)
 	data_result := utils.RespLoadJson[SearchJson](body).Data
 	for index, value := range data_result {
 		fmt.Println(strconv.Itoa(index+1) + "# " + value.IP)
